move calculate_settling_rate into grain_hill
