+++ grainhill/__init__.py
@@ -1,6 +1,7 @@
 from .grain_hill import GrainHill
 from .bmi_grain_hill import BmiGrainHill
 from .grain_hill import plot_hill
+from .grain_hill import calculate_settling_rate
 from .grain_hill import CTSModel
 from .block_hill import BlockHill
 from .grain_facet_model import GrainFacetSimulator
