+++ grainhill/grain_facet_model.py
@@ -4,7 +4,7 @@ Model of normal-fault facet evolution using CTS lattice grain approach.
 """
 
 import sys
-from grainhill import CTSModel, plot_hill
+from grainhill import CTSModel, plot_hill, calculate_settling_rate
 from grainhill.lattice_grain import (lattice_grain_node_states,
                                      lattice_grain_transition_list)
 import time
@@ -376,31 +376,6 @@ def get_params_from_input_file(filename):
     mpd_params = load_params(filename)
 
     return mpd_params
-
-def calculate_settling_rate(cell_width, grav_accel):
-    """
-    Calculate and store gravitational settling rate constant, based on
-    given cell size and gravitational acceleration.
-
-    Parameters
-    ----------
-    cell_width : float
-        Width of cells, m
-    grav_accel : float
-        Gravitational acceleration, m/s^2
-
-    Notes
-    -----
-    Returns settling rate in yr^-1, with the conversion from s to yr
-    calculated using 1 year = 365.25 days.
-
-    Examples
-    --------
-    >>> round(calculate_settling_rate(1.0, 9.8))
-    69855725.0
-    """
-    time_to_settle_one_cell = np.sqrt(2.0 * cell_width / grav_accel)
-    return SECONDS_PER_YEAR / time_to_settle_one_cell
 
 def main(params):
     """Initialize model with dict of params then run it."""
